crossroad-mingw-install: directory creation may override files.
I had the strange case where a directory created by an install existed
as a file, which raised an exception.
Not sure this is the right solution, but I will just output a warning
and delete it. In the end, we are still working in a self-contained
cross-compilation prefix, so I assume that whatever we delete is not a
huge loss.