crossroad: add $BUILD variable to bashrc.
If --host is set but not --build, autoconf attempts a run test before
assuming it is a cross-compile environment.
See doc, or more simply check the source doc:
http://www.gnu.org/software/autoconf/manual/autoconf.html#Hosts-and-Cross_002dCompilation
This is acceptable behavior unless your environment is able to run some
basic Windows binaries, which now happens with Wine integration in Linux
systems.  I opened a ticket with a patch for this specific case:
https://savannah.gnu.org/support/index.php?108405
But just in case it is not accepted for some compatibility reason, for
instance, or simply for older autoconf versions, we want to provide
users an easy $BUILD variable. Since this variable is kind of hard to
guess, I copied `config.guess` script straight from autoconf master
repository.