setup.py: fixes pip install.
This fixes "option --single-version-externally-managed not recognized"
error when installing with `pip`. It seems that pip would assume any
package uses setuptool.
So I use setuptools instead if it is already loaded (which is the case
when pip installs). Additionally if the user set USE_SETUPTOOLS env
variable, we also force setuptools.
This little trick is borrowed from SaltStack code, a Apache v2 software.