crossroad-mingw-install: rpm2cpio preferred to 7z and better error reporting.
If rpm2cpio is available, use it instead of 7z. It prevents the
intermediary cpio file step.
Also now I have better error reporting. First it checks for the various
command dependencies and stop if any missing, *before* trying to
install. Second an extraction error is now a stopper and provoke
immediate failure.