Add support for builds with mix of native and cross-compilation.
Some project would build native tools to be used in the rest of the
compilation (which may be native too, or cross-compiled). I found this
case on GTK+3 (master) and fontconfig master. Usually there would be
some *_FOR_BUILD macro in the configure.ac of an autotools build (for
instance AX_CC_FOR_BUILD to search for a native compiler).
* When setting pkg-config variables to search in the cross-build path,
it made it impossible to search native libraries anymore. So instead
I create a separate pkg-config for each supported cross-environment,
so that we can call both the native and cross pkg-config.
* Same for gcc. Setting CFLAGS and LDFLAGS was a bad idea because they
are use everywhere by the autotools (some project use CFLAGS_FOR_BUILD
but that's apparently not a standard). And a globale CPATH is also a bad
idea because used by any call of gcc (cross-compilation or native one).
Thus I create a layer gcc which sets the right env variable CPATH and
LIBRARY_PATH only for the cross-comp gcc).
Also it's probably better to leave the usage of CFLAGS and LDFLAGS to
the user, for instance to add some optimization flags, debugging,
platform-specific flags, etc.