Replace LIBRARY_PATH by -Lsomelib flags directly on the command line.
It seems that LIBRARY_PATH would not be used when cross-compiling since
it could mess up with the native compiler.
I read this claim on MinGW doc:
http://www.mingw.org/wiki/HOWTO_Specify_the_Location_of_Libraries_for_use_with_MinGW
"The GCC manual states that this technique may be used only when GCC is
configured as a native compiler; thus, this method of specifying the
supplementary library path may not be used when your MinGW GCC is
configured as a cross-compiler."
Unfortunately I could not locate confirmation of this in gcc manual, but
it indeed does not seem to work.
There may be another "specs file" method, but this one works ok too.
Note that it should normally be the same for CPATH, but for some
unknown reason, the include dirs set in CPATH are still used during a
cross-build.