wine: adds a WINEPREFIX env variable.
This allows to install things directly in the prefix. Some projects are
hard to crossbuild, for instance python which even has specific test
conditions in the build to prevent cross-compilation with mingw.
But they can be installed with wine (I could run only the 32-bit
installer though).
There is still the problem that the PATH won't be good to find software
installed this way. Also this should be specific to Windows crossbuild.
But that's WIP.