setup: install the manual under $prefix/share/man.
The Filesystem Hierarchy Standard is kind of inconsistent about it
since it says that /usr/share/man is the right directory for /usr,
but:
"Manual pages for commands and data under /usr/local are stored in
/usr/local/man. Manual pages for X11R6 are stored in /usr/X11R6/man."
http://www.pathname.com/fhs/pub/fhs-2.3.html#USRSHAREMANMANUALPAGES
So what for man pages of a given custom prefix?
It turns out most software seem to install their man under
$prefix/share/man so I'll do the same. Also there is supposed to be
a locale hierarchy but once again most user software don't follow it,
and when I tried, man was unable to find it even when installed in my
locale. So I just follow and drop the locale subdirectory.