Add very basic scons support.
Scons has no real support for cross-compilation, as far as I could find.
There seems to be some common usage though, and the following added
support in crossroad allowed me to build libmypaint (though only as
static lib, which was useless to me).
Let's commit this, though not sure how usefull it is in the end.