setup: install the bash-completion script under a more standard path.
$PREFIX/share/bash-completion/completions/ is the standard path and if
we happen to use a common prefix, the completion script may even end up
sourced by default by the system. Complete a bit the manual for other
alternative to source this file (for instance the bashrc or
~/.bash_completion). Checking the contents of
`/usr/share/bash-completion/bash_completion` is actually the main and
simplest reference about where and how to setup a bash-completion script
for a program.