Do not set a CROSSROAD_PREFIX as the --prefix in Android builds.
This is usable for Windows builds only where the installation path is
unknown anyway as is common on such platform.
Android is back to being a nearly-normal Linux with hard-set
installation prefixes. Not sure yet if that should be /usr/, /usr/local/
or something else completely, I will have to check what are the common
plaform paths for third-party applications. But for now, let's just use
/usr/local/ and implement the logics of an intermediate installation
directory using DESTDIR environment variable.