platforms: set CMAKE_CROSSCOMPILING_EMULATOR for Windows crossbuilds.
This will make TRY_RUN() configuration works with Wine, and tests work
as well. TRY_RUN() should really not be used during configuration IMO,
or it should be protected. Yet someone reported to me a program (CGAL)
which was using it and cross-compilation was failing because of it. I
was able to compile the version 4.9 by using Wine in configuration.
Since Wine is not 100% reliable though, there should be a way to disable
this variable. Something to improve later!