platforms: fix toolchain host.
Docs developer.android.com says the toolchain name is: x86-<gcc-version>
Yet after installing it, I notice it is actually i686-linux-android-*
(which, by the way, is a much better naming!). Just fix this.
There is still a bit of inconsistency in that the NDK script
(make-standalone-toolchain.sh) won't recognize 'i686-linux-android' in
the --toolchain option, so I need to add a new hosts dictionnary, to
keep track of the toolchain name vs the actual host name.