platforms: more files needed to be fixed as executable.
The files in bin/ are not enough and I had build issues. I compared with
the original archive, checking with the following command:
$ find android-24-arm -executable -type f
It looks like making executable anything under a bin/ or libexec/
directory, as well as python files (not sure if these are necessary,
since when used as modules, or prefixing standalone calls with the
interpreter, it's unneeded, but some were executable; whatever, better
safe than sorry) should do the trick.