platforms: set the __ANDROID_API__ macro for Android builds.
This was used in a few headers shipped with the NDK. I would assume that
the compiler would set it appropriately since it was generated knowing
the API version, but it seems no. Various programs were failing to
build, in particuler under API 23, with undefined references to stderr
and stdout at link time. Just setting it to the right API version fixed
the build. Let's just always set this macro in the preprocessor flags.