Add "native" platform.
This seems like a weird idea but it actually allows to easily install
software in their own custom prefixes, with PATH and LD_LIBRARY_PATH in
particular updated. This also allows to easily build and install several
versions of a software (for instance GIMP master and 2.8) without them
interfering with one another.