scripts: append to local $LD_LIBRARY_PATH for native builds.
We have to assume a user-set LD_LIBRARY_PATH variable was made for the
native platform. Of course, for cross-builds, one could think this
variable is useless anyway. Yet with various emulators, it is possible
to execute non-native executables on Linux. And when this happens,
LD_LIBRARY_PATH may be used as well.