platforms: set CMAKE_SYSTEM_PROCESSOR for Win32 platforms.
CMake wiki indicates that this is not set automatically when
cross-compiling: https://cmake.org/Wiki/CMake_Cross_Compiling
Yet this can be necessary on some CMake scripts (I use it in uchardet
for instance).
CMake docs also indicates that this value would be set to the
PROCESSOR_ARCHITECTURE environment variable in Windows:
https://cmake.org/cmake/help/v2.8.8/cmake.html#variable:CMAKE_SYSTEM_PROCESSOR
Finally I could find that this variable usually takes "x86" or "AMD64"
values for 32 and 64-bit targets respectively. So I'll just reproduce
these values:
https://blogs.msdn.microsoft.com/david.wang/2006/03/27/howto-detect-process-bitness/