scripts: pkgconf does not like empty PKG_CONFIG_LIBDIR.
To be accurate, it would seem like it considers it the same as being
unset. Well it's a bit unclear since it looks it looks inside
PKG_CONFIG_PATH list in priority (even though that's considered the
secondary source), but fallbacks to the system dirs (which is just wrong
for cross-compilation, obviously!).
Anyway just set PKG_CONFIG_LIBDIR to the prefix paths to prevent this
unwanted behavior. The alternative paths within PKG_CONFIG_PATH will
still be looked too.

I believe it to be one more difference with original pkg-config as I
know it used to work fine by setting to empty string.