Use Fedora repository for Mingw packages.
This is experimental as it is currently a lot slower. My guess is that
this is because there are no dedicated repo (so we use main repositories
and have to extract only Mingw packages for a given arch), so it would
make parsing much longer. I will have to do some optimization, but this
is a first quick, dirty and working commit.

Also Suse repo are still available with the new CLI option --arch to
crossroad-mingw-install.py. The change is mostly due to less outdated
versions of packages in the Fedora repository.