Serialize package data after first read.
As expected the slowest part with Fedora repository was parsing since
they are so huge, and the MinGW packages are only a small piece of the
whole list.
So let's serialize and unserialize in Python binary format, which makes
at least parsing extra fast except for the first time each time the repo
updates.