scripts: check for specific distrib lib path.
So it turns out that on some distributions (e.g. Debian and Ubuntu),
libraries are not necessarily in lib/, lib64/ or similar but in
lib/x86_64-linux-gnu/ (for instance, in the common x86_64 linux case).
Such path base need to be added to LD_LIBRARY_PATH, GI_TYPELIB_PATH and
PKG_CONFIG_PATH. I put them in higher priorities to the generic lib/ or
lib${WORD_SIZE}/ case.