src: run the custom bashrc with -c when crossroad called with --run…
… and when exiting immediately after the script.

For some reason, crossroad was immediately returning (without actually
running the install script) when bash was run with --rcfile instead. It
seemed like the CI was able to say when a command was meant to be
interactive (that is an assumption based on the fact it was also
breaking immediately out of other interactive commands such as dnf
without --assumeyes), even though in this particular case, it was not
actually interactive because the start script was forcing bash to exit
immediately after running.
Anyway now special-case this particular case to a really non-interactive
call, and it looks like now crossroad is perfectly usable as a gitlab-ci
build script.