scripts: add a `ccd` environment command to cd into $CROSSROAD_HOME.
If I want to create build directories for my projects, we now have the
$CROSSROAD_HOME concept (since commit bd09af6b). This is different from
the #CROSSROAD_PREFIX as these are not released/installed files, but
more build-time files.
Unfortunately it was a bit cumbersome to use. So I create the `ccd`
command (for "Crossroad cd") to change directory inside Crossroad home.
Additionally if a folder does not already exist, it will propose to
create it (hence acting as a `mkdir`). This makes navigating between a
crossroad project's specific files and the normal home files a lot
easier.