platforms, scripts: add a `source` subcommand for the Win32 platforms.
This allows to select the source for the packages. By default, if you
are using a distribution which has cross-build packages (SUSE and Fedora
as far as I know, so far), this will be chosen since you will avoid most
inconsistencies this way.
But if you are using a distribution which provides cross-compilers but
no pre-build packages (such as Debian derivatives, unless mistaken), you
may want to try out different sources for one which will work better
with your available cross-compiler/stdlib.

New commands are:
- crossroad source -> list available sources.
- crossroad source --set=suse -> set to "suse" source for instance.