scripts: serialize the Arch repository data too.
And especially don't even check if a new version of the repository has
been published, unlike RPM repositories. This is because RPM repos
provide such info with a very small file to download only.
For the Arch repository, we'd basically have to redownload the whole DB
(just a few MiB, but still too much).