scripts: (Win32|64) add packager URL and package URL in `info` display.
The package URL was not displayed by the `info` command until now.
Useful to check were a package is taken from.
As for the packager URL, it can be whatever web page is used by upstream
packagers to display information about their work. For Msys2 in
particular, it will be the package page on the packages.msys2.org
website. No such links is implemented for now for the RPM repository
(not sure if they have such information page, but if not maybe I could
show the RPM specs file).