scripts: single equality sign is more portable.
Basically, according to my searches, double equality '==' is an alias to
'=', but it is not always implemented (unlike '='). So we should always
use single '='.

Alternatively, using [[]] test syntax works also as '==' apparently
always exists with this syntax. This was the initial fix proposed by
Yannick Le Roux (thanks!) who encountered issues with zsh. Yet [[]] is
not POSIX, hence may not exist in some shells (though in practice I saw
this in many shells, but I don't know them all!). So let's just go with
[] and single '='.