Release crossroad version 0.9.0.
Could also be named the "GIMP release" as it benefited from fixes to
various issues discovered while setting up Windows 32 and 64-bit builds
for GIMP on our Gitlab continuous integration platform. This crossbuild
has run with success since September 2019 (slowly improved here and
there since then) and made nice automatic/nightly test builds for
Windows.
I also implemented CI builds for Siril (an image processing software for
astronomy) which also works well and is now used by the project to
produce nice and ready-to-use packages.

So we can safely say that Crossroad is getting really stable and got its
trial by fire now (even though for me, it has been the case for years,
as I've used it daily ever since I added a native target, so 3 years of
daily usage now; but that's confirmation in the wild now and for the
cross-compilation use case in particular).