scripts: better error when crossroad-gcc wrapper fails.
As we install our GCC wrapper, it is found by the build system, but if
the real compiler is not behind, actual compilation will obviously fail.

On a CMake case, we had such error:
> /home/jehan/.local/share/crossroad/bin/x86_64-w64-mingw32-g++: line 135: -o: command not found

This is not very explicit. Now the error message will be:

> ** [Crossroad] Error: actual compiler (x86_64-w64-mingw32-g++) does not exist in your PATH.
> ** [Crossroad] Error: please install x86_64-w64-mingw32-g++.

Hopefully people will better understand that they need to install the
cross-compiler.