scripts: allow using official MSYS2 mirror when main repo is down.
Thanks to the main MSYS2 repository being down today, I discover their
officially curated list of mirrors. It is to be noted that this list is
also available as a package there:
https://packages.msys2.org/package/pacman-mirrors
But if I were to download it as a package from the main repository, then
it would make no sense the day where this repo is down (like today)
because Crossroad would not be able to download it either. Well it could
still be useful if the mirror list is kept, for long-use development
environment (like personal developer builds), but not for one-use
builds, like in a CI.

Optionally I could save a copy of the list with Crossroad itself, but
then it would not be updated often enough (rather than missing new
mirrors, it could become a security issue if a mirror is removed from
the official list, but Crossroad still were to have it).

So I chose to download the list from the official repository of this
package on Github as it's a different source (hence less chance that
both Github and the main repository are down the same day, though not
impossible of course). Note that I only try to download it when a
download error occurs.
Also I decrease the timeout and don't retry a same URL twice anymore on
MSYS2 because instead, I retry on mirrors (but I still keep longer
timeouts and more retries on same URL for Fedora and OpenSUSE
repositories).