platforms: copy all system DLLs for Windows targets.
This fixes 2 issues:
(1) Since commit b70b446, Crossroad was symlinking the Windows DLLs,
which was enough for using the prefix from a shared folder used in a VM
for instance, or to make an archive (symlinks would be dereferenced
automatically and real file copied). But I encountered an issue in GIMP
CI with the 32-bit Windows build which was missing a libgcc DLL. The
file was there, but as a broken symlink, with no original anymore. Let's
just copy the real one. Even if not needed in the end, we have the tool
to inspect binaries and only keep used libs.
(2) I realize that `gcc -print-file-name=` is where system DLLs are
installed in Debian, but on Fedora, there are in a subdir of `gcc
-print-sysroot` (which returns an empty string on Debian). So let's just
check both and copy over all DLLs found in any with these directories.