src: exit and output an explicit error on stderr when bashrc or zsh…
… files are missing
This issue happened with the older pip3 issue (in version 18.1 at
least), which I wrote about in the README a few days ago. It would start
the Crossroad environment but fail to run the startup files containing
all sort of needed variables. And there was no error or warning being
displayed (other than the prompt would not be changed, and Crossroad
inner commands would not work).

Now Crossroad will properly exit directly with an error, which will be a
much nicer way to handle the issue (we'll know directly the problem).