scripts: reset shell color output with \033[0m
On my local shell (bash) it works well with "\033[00m" (double 0, not
sure where it came from) or with simple 0. On Gitlab, the red color
seems to stick, not sure why. This makes the whole build output red ever
since crossroad loads which makes things a bit harder to debug.
It may not be this reason, but anyway, let's try.