src: the --depends option also starts the environment.
Without this, when you want to create a new environment with a deps, you
do it in too many steps: first create the env normally, then exit it,
then add the dependency, then get back in.

Now it's all in one step. If really you were in a rare case where the
goal was not to get in immediately, you can get out easily. It's more
often a win.