scripts: RPM files may now uncompress as .cpio.zstd.
This is what I realized when testing new Fedora repositories.

I'm also adding a small warning that `rpm2cpio` is anyway a recommended
tool as it handles everything. I had one case of package (binutils),
which for some reason 7z failed to extract with "Headers Error" and a
"There are data after the end of archive" warning.